# CI performance improvement (task #11384)