--- a/.travis.yml
+++ b/.travis.yml
@@ -38,7 +38,7 @@ install:
   - ./bin/composer install --no-interaction --no-progress --no-suggest
   - yarn install
   - mkdir -p build/test-coverage build/test-results
-  - ./bin/build app:install CHOWN_USER=$USER,CHGRP_GROUP=$USER,DB_NAME=app,DB_ADMIN_USER=root,DB_USER=root --quick
+  - ./bin/build app:install CHOWN_USER=$USER,CHGRP_GROUP=$USER,DB_NAME=app,DB_ADMIN_USER=root,DB_USER=root --skip-test-db
 
 before_script:
   - ./bin/phpserv >/dev/null 2>&1 &