Currently, during a project installation/update, we first execute the CakePHP migrations against the test database to catch any errors before we run them against the live database.

This PR introduces a flag which is used to skip running migrations against the test database on CI builds since each build is run against a new environment without any pre-existing database data.

Listed below are some benchmarks run locally, executed three times for each case. The difference will dramatically increase in child projects where the number of tables is much greater.

Without `--quick` flag:
- 18.177s
- 19.495s
- 18.864s

With `--quick` flag:
- 11.243s
- 10.720s
- 11.139s