@softius 2019-05-13
@georgeconstantinou let's change the option name to something more specific like `--skip-test-db`.