@georgeconstantinou 2019-05-13
> @georgeconstantinou let's change the option name to something more specific like `--skip-test-db`.

@softius done https://github.com/QoboLtd/project-template-cakephp/pull/793/commits/c5c1c76593880e291f3712d5b2cf418f87a84a15